Make lexer output proper Bison location info
This might introduce some minor changes in the grammar, as we have to change
the way we lex whitespace in order to count lines, and the result is linebreaks
aren't valid everywhere they used to be. But the whole android code base
builds, and in coming changes we'll be able to put some of that back (we'll be
reducing complexity in the lexer and moving that in to the grammar).

Change-Id: I6a688b2dc45e6ea4838af466d2cca5e1b26f916c
Signed-off-by: Casey Dahlin <sadmac@google.com>