Simplify the logic in compile_aidl_to_java()
1) Explicitly check that the input .aidl file has exactly one
   declared item, and that it is an interface.
2) Use a few C++isms to simplify some loop structure.
3) Reduce levels of nesting in loops by continueing on errors.

The code may have at some point supported declaring parcelables next to
intefaces, but later assertions (file names match declared classes) make
this impossible in practice.  Since it simplifies the code to do so,
make the check explicit.

The build system always passes -b (for break on parcelable) when
compiling aidl to java (as opposed to preprocessing), so this is how the
tool is used in practice.

Test: unittests pass, clean build of aosp_arm works
Bug: 23599760

Change-Id: If291cdd9d02a25df0be08c04a45ba379be7d0969