Don't segfault on missing package statement
.aidl files that are missing a package statement record this as an empty
c string.  When we went to look up the type of the generated Java class
we would segfault when a lookup returned a nullptr instead of the
type representing the class.  This turned out to be because we were
looking for ".IType" rather than the expected "IType".

Test: Compile a trivial .aidl file that defines an interface but no
package.  Observe that this now generates code rather than segfaulting
without a message.

Change-Id: Ibe4c001e412c73fea2a7acafbc91e83d4ab929a0