Respect aidl's -b flag again
This flag indicates that we should (or not) report errors on
encountering a parcelable in an input .aidl file.  In the platform
build we always break on parcelables, since we only get called with
interface definitions.

We need to obey this in order to not break Android studio's usage.
There, developers expect to put *all* .aidl files into aidl and have it
only generate code where appropriate.

Bug: 25501507
Test: unit tests

Change-Id: I91df9b0ef65d4d063a7a898a5948d0429af3e255