Make TypeNamespace take AidlNode subclass arguments
Because annotations are attached to AidlType, we need to know the
complete AidlType of arguments and return values in order to decide
which generated type to use.  Make it an invariant that a TypeNamespace
takes only subclasses of AidlNode from aidl.cpp to make this easier to
reason about.

For now, leave most of the existing implementation in place, but push it
down into LanguageTypeNamespace.

Bug: 26729450
Test: unittests still pass

Change-Id: I6684564995c734def96daeccc03dfd5d264469ee